app/backend/api/pyproject.toml
[tool.poetry]
name = "azure101sample"
version = "0.1.0"
description = ""
authors = ["kohei3110 <kk.31108424@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
apscheduler = "^3.11.0"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-exporter-otlp-proto-http = "^1.29.0"
azure-monitor-opentelemetry-exporter = "^1.0.0b33"
aiofiles = "^24.1.0"
azure-ai-projects = "^1.0.0b5"
azure-identity = "^1.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
